module6:
- changing variables names in the Lesson 01 - lessons and challenges:

-- instead of replacing AGE_PERCENTIL directly in the cleaned dataset - with .astype('category').cat.codes, I created a new dataframe that is a copy of cleaned one and I updated AGE_PERCENTIL with category type.

I did that because I needed the original AGE_PERCENTIL values to apply a different technique to categorize the values.

So I created two diferents dataframes from the cleaned one which each one used a different technique to convert the values in AGE_PERCENTIL variable.

In this way, I can train and test models with these dataframes and then check what was the best technique to update the variable.

-- corrected the grammatical error in the ENCONDED word

- lesson 02: done
- challenges lesson 2: done